Remove permutation importance
This commit removes the `sf.stats.permute_importance` function (and
related functions), and the `permutation_importance` importance
argument to `Project.evaluate()`.

The `sf.stats.permute_importance` function was not fully
implemented/tested, and it was not implemented in the PyTorch backend.
As this half-implemented function was not being actively used,
it has now been removed to decrease maintenance burden.