Handle missing pyvips or libvips (#169)
In some cases, someone may wish to use Slideflow without slide-processing (such as if TFRecords are already generated, or to use other functionality). Presently, however, Slideflow cannot be imported without libvips/pyvips installed.

This change will log an error if pyvips/libvips is not installed, but will allow the module to be imported.