Log warning if Dataset cannot find a config source
Currently, if a dataset source is supplied to `sf.Dataset` (either manually or via project `settings.json`) and that source does not exist in the dataset configuration JSON file, the missing source will be quietly ignored.

This adds a warning message that a source could not be found.

Fixes SLID-61