Tile extraction bug fix & testing improvements
- Fix bug where, in certain environments,
tile extraction would hang if started after
pyvips was imported and used in another
process. This was due to the type of shared
multiprocessing pool used during tile
extraction (fork). The multiprocessing method
was switched to "spawn", which fixes the bug.
This slightly delays the startup time for tile
extraction by a few seconds.
- Add `slide-threads argument to `test.py`,
which specifies number of threads to use when testing
tile extraction.  Setting the number of threads to
1 circumvents multiprocessing and may help with debugging.
- Add `unit` as argument to `test.py` to disable unit testing.
- Tile extraction with `num_threads=1` does not use any pools