Support calculating features from Imagnet models
- Features (layer activations) can now be generated from Imagenet-pretrained (rather than only Slideflow-trained) models, by passing a loaded Tensorflow/Keras or Torch model to `DatasetFeatures` (rather than the path to a saved model).
`Project.generate_features_for_clam()` now accepts the name of an architecture to the `model` argument, so features can be calculated from a pretrained (but not Slideflow-trained and saved) model. If providing a model architecture name rather than the path to a saved model, a Dataset must be explicitly passed to `dataset`. E.g.: `Project.generate_features_for_clam(model='xception', dataset=...)`
- Improved CLAM error function when there are insufficient tiles per slide (min_tiles=8)