Changes to `Project.ensemble_train_predictions()`, round 1
This first round of changes to the `Project.ensemble_train_predictions()` includes a couple of stylistic improvements that improve readability.

First, the variable `ensemble_dirs` was renamed to `member_paths`. This was done for two reasons:

1) "Member" is slightly clearer, since in this case we are referring to a member of the ensemble
2) For consistency, I try to refer to folder names as 'dir' and full paths as 'path'.

Next, when iterating through a list in Python, it is generally better to iteration using the syntax:

```python
for item in item_list:
  item.do_something()
```

Rather than:

```python
for i in range(len(item_list)):
  item_list[i].do_something()
```

... as the former is more readable. If you need the index number for something (which is not needed in this case), you can use `enumerate:`

```python
for i, item in enumerate(item_list):
  ...
```

Third, the `os.path.join()` function is used to join directories together as it is safer than manual string editing.