TFRecord index files without ZIP archive format (#249)
On databricks with dbfs file storage, *.zip files are not supported. TFRecord index files are written by default as *zip files (using numpy.savez). This PR adds ability to save the index files as non-zip numpy arrays (using numpy.save), if the environmental variable SF_ALLOW_ZIP is set to 0. This needs to be controlled via environmental variable to ensure the setting is propagated to all child processes in multiprocessing pools.

See #245